+++ src/main/java/gg/packetloss/grindstone/economic/lottery/mysql/LotteryWinnerStatement.java
@@ -8,6 +8,7 @@ package gg.packetloss.grindstone.economic.lottery.mysql;
 
 import gg.packetloss.grindstone.data.MySQLPreparedStatement;
 
+import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
@@ -17,9 +18,9 @@ public class LotteryWinnerStatement implements MySQLPreparedStatement {
     private Connection con;
     private final long date;
     private final int player;
-    private final double amount;
+    private final BigDecimal amount;
 
-    public LotteryWinnerStatement(int player, double amount) {
+    public LotteryWinnerStatement(int player, BigDecimal amount) {
         this.date = System.currentTimeMillis() / 1000;
         this.player = player;
         this.amount = amount;
@@ -37,7 +38,7 @@ public class LotteryWinnerStatement implements MySQLPreparedStatement {
         try (PreparedStatement statement = con.prepareStatement(sql)) {
             statement.setLong(1, date);
             statement.setInt(2, player);
-            statement.setDouble(3, amount);
+            statement.setBigDecimal(3, amount);
             statement.execute();
         }
     }
